implement input select component (#2)
* feat: publishing the Input part of the InputSelect component

* feat: Modify Dropdown to select options according to key input (ArrowUp, ArrowDown, Enter)

* feat: Close OptionList when click outside

* feat: Close OptionList when click outside

* fix: fix an issue where an object string was output when an option was an object

+++ packages/shared-ui/src/components/Icon/CheckIcon.tsx
@@ -50,7 +50,7 @@ const getCheckIcon = (valiant?: Valiant, color: ColorMap = ColorMap.green) => {
           />
         </svg>
       );
-    case 'single':
+    case 'double':
       return (
         <svg viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
           <path
